LPS-26321 We already ensure that PluginContextListener is the FIRST listener that is initialized by the servlet container, because it's the first listener. But the servlet container construts all the listeners ahead of time. So this means someone can do an attack when the listener is contructed. This commit moves the list of listeners to a context init param, that is then constructed and initialized by the portal AFTER security is enabled.
Also, the servlet is constructed by the wrong class loader. It's the web app's parent class loader, meaning, it's not registered with PACL, meaning, it's not protected. So we have to wrap the servlet. It also needs to be part of the portal lifecycle so that we call destroy on it BEFORE the portal is shut down. Just in case someone tries to attack during construction, init, or destroy phases.